DPO-896 pk int -> long

+++ AvoinDataCommon/src/main/java/fi/livi/rata/avoindata/common/dao/trackwork/TrackWorkNotificationIdAndVersion.java
@@ -5,10 +5,10 @@ import io.swagger.annotations.ApiModelProperty;
 public interface TrackWorkNotificationIdAndVersion {
 
     @ApiModelProperty("Track work notification id")
-    Integer getId();
+    Long getId();
 
     @ApiModelProperty("Track work notification version")
-    Integer getVersion();
+    Long getVersion();
 
 }
