+++ AvoinDataCommon/src/main/java/fi/livi/rata/avoindata/common/dao/trackwork/TrackWorkNotificationRepository.java
@@ -16,13 +16,13 @@ import java.util.Set;
 public interface TrackWorkNotificationRepository extends CustomGeneralRepository<TrackWorkNotification, TrackWorkNotification.TrackWorkNotificationId> {
 
     @Query("SELECT t.id.id AS id, t.id.version AS version FROM TrackWorkNotification t WHERE t.id.id IN (:ids) ORDER by id, version ASC")
-    List<TrackWorkNotificationIdAndVersion> findIdsAndVersions(@Param("ids") Set<Integer> ids);
+    List<TrackWorkNotificationIdAndVersion> findIdsAndVersions(@Param("ids") Set<Long> ids);
 
     @Query("SELECT t FROM TrackWorkNotification t WHERE t.id.id = :id ORDER by id, version ASC")
-    List<TrackWorkNotification> findByTwnId(@Param("id") int id);
+    List<TrackWorkNotification> findByTwnId(@Param("id") long id);
 
     @Query("SELECT t FROM TrackWorkNotification t WHERE t.id.id = :id AND t.id.version = :version")
-    Optional<TrackWorkNotification> findByTwnIdAndVersion(@Param("id") int id, @Param("version") int version);
+    Optional<TrackWorkNotification> findByTwnIdAndVersion(@Param("id") long id, @Param("version") long version);
 
     @Query("SELECT t.id.id AS id, MAX(t.id.version) AS version FROM TrackWorkNotification t WHERE t.modified BETWEEN :start AND :end GROUP BY t.id.id ORDER BY id ASC")
     List<TrackWorkNotificationIdAndVersion> findByModifiedBetween(@Param("start") ZonedDateTime start, @Param("end") ZonedDateTime end, Pageable pageable);
